+++ Lib/include/Utilities/Smooth.h
@@ -97,11 +97,11 @@ inline Array<double> Smooth::gaussian(
 	normalization = 1/normalization;
 
 	Array<double> result({data.getRanges()[0]}, 0);
-	for(int n = 0; n < data.getRanges()[0]; n++){
+	for(unsigned int n = 0; n < data.getRanges()[0]; n++){
 		for(
-			int c = std::max(0, n - (int)windowSize/2);
+			int c = std::max(0, (int)n - (int)windowSize/2);
 			c < std::min(
-				n + (int)windowSize/2 + 1,
+				(int)n + (int)windowSize/2 + 1,
 				(int)data.getRanges()[0]
 			);
 			c++
@@ -142,11 +142,14 @@ inline std::vector<double> Smooth::gaussian(
 	normalization = 1/normalization;
 
 	std::vector<double> result;
-	for(int n = 0; n < data.size(); n++){
+	for(unsigned int n = 0; n < data.size(); n++){
 		result.push_back(0);
 		for(
-			int c = std::max(0, n - (int)windowSize/2);
-			c < std::min(n + (int)windowSize/2 + 1, (int)data.size());
+			int c = std::max(0, (int)n - (int)windowSize/2);
+			c < std::min(
+				(int)n + (int)windowSize/2 + 1,
+				(int)data.size()
+			);
 			c++
 		){
 			result.at(n) += data.at(c)*exp(-(c-n)*(c-n)/(2*sigma*sigma));
